# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Learning Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with phone number and PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register-pin:
    post:
      operationId: registerPin
      tags: [auth]
      summary: Register a PIN for a phone number (first-time setup)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterPinInput"
      responses:
        "201":
          description: PIN registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/create-account:
    post:
      operationId: createAccount
      tags: [auth]
      summary: Create a new account — sends a temporary PIN by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountInput"
      responses:
        "201":
          description: Account created, temporary PIN sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateAccountResult"
        "400":
          description: Invalid input or email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-temp-pin:
    post:
      operationId: verifyTempPin
      tags: [auth]
      summary: Verify the temporary PIN sent by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyTempPinInput"
      responses:
        "200":
          description: PIN verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyTempPinResult"
        "400":
          description: Invalid or expired PIN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/set-pin:
    post:
      operationId: setPin
      tags: [auth]
      summary: Set a permanent PIN after temp PIN verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPinInput"
      responses:
        "200":
          description: PIN set, account activated, user logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Invalid request or PINs do not match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/resend-pin:
    post:
      operationId: resendPin
      tags: [auth]
      summary: Resend the temporary PIN email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendPinInput"
      responses:
        "200":
          description: PIN resent successfully
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    type: boolean
        "400":
          description: Account not found or already active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── ZONES ───────────────────────────────────────────────────────────────────
...
  /coordinators:
    get:
      operationId: listCoordinators
      tags: [coordinators]
      summary: List all zonal coordinators
      responses:
        "200":
          description: List of coordinators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coordinator"
    post:
      operationId: createCoordinator
      tags: [coordinators]
      summary: Create a new zonal coordinator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoordinatorInput"
      responses:
        "201":
          description: Coordinator created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coordinator"
  /coordinators/{id}:
    get:
      operationId: getCoordinator
      tags: [coordinators]
      summary: Get coordinator by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Coordinator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coordinator"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /coordinators/{id}/stats:
    get:
      operationId: getCoordinatorStats
      tags: [coordinators]
      summary: Get coordinator performance stats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Coordinator stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoordinatorStats"

  # ── TRAINERS ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # AUTH
    LoginInput:
      type: object
      required: [pin]
      properties:
        phone:
          type: string
          description: Indian phone number (10 digits) — legacy phone login
        email:
          type: string
          description: Email address — for email-based login
        pin:
          type: string
          description: 4-digit PIN
    RegisterPinInput:
      type: object
      required: [phone, pin, role]
      properties:
        phone:
          type: string
        pin:
          type: string
        role:
          type: string
          enum: [management, coordinator, trainer, student]
        entityId:
          type: ["integer", "null"]
          description: ID of the trainer/coordinator/student this user maps to
    AuthResult:
      type: object
      required: [success, role, userId]
      properties:
        success:
          type: boolean
        role:
          type: string
          enum: [management, coordinator, trainer, student]
        userId:
          type: integer
        entityId:
          type: ["integer", "null"]
          description: ID of the linked entity (trainer, coordinator, student)
        name:
          type: ["string", "null"]

    # ZONE
...
    Coordinator:
      type: object
      required: [id, name, zoneId, zoneName]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        zoneId:
          type: integer
        zoneName:
          type: string
    CoordinatorInput:
      type: object
      required: [name, zoneId]
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        zoneId:
          type: integer
    CoordinatorStats:
      type: object
      required: [coordinatorId, name, zoneName, centreCount, trainerCount, studentCount, completionRate]
      properties:
        coordinatorId:
          type: integer
        name:
          type: string
        zoneName:
          type: string
        centreCount:
          type: integer
        trainerCount:
          type: integer
        studentCount:
          type: integer
        completionRate:
          type: number

    # TRAINER
...
    CreateAccountInput:
      type: object
      required: [email, phone, role]
      properties:
        email:
          type: string
          format: email
        phone:
          type: string
          description: Indian phone number (10 digits)
        role:
          type: string
          enum: [management, coordinator, trainer, student]
    CreateAccountResult:
      type: object
      required: [success, email]
      properties:
        success:
          type: boolean
        email:
          type: string
    VerifyTempPinInput:
      type: object
      required: [email, pin]
      properties:
        email:
          type: string
        pin:
          type: string
          description: 4-digit temporary PIN received by email
    VerifyTempPinResult:
      type: object
      required: [success, email]
      properties:
        success:
          type: boolean
        email:
          type: string
    SetPinInput:
      type: object
      required: [email, newPin, confirmPin]
      properties:
        email:
          type: string
        newPin:
          type: string
          description: New 4-digit permanent PIN
        confirmPin:
          type: string
          description: Must match newPin
    ResendPinInput:
      type: object
      required: [email]
      properties:
        email:
          type: string

    # USER (internal)
...