# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Learning Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthResultRole } from './authResultRole';

export interface AuthResult {
  success: boolean;
  role: AuthResultRole;
  userId: number;
  /**
     * ID of the linked entity (trainer, coordinator, student)
     * @nullable
     */
  entityId?: number | null;
  /** @nullable */
  name?: string | null;
}