# lib/api-zod/src/generated/types/coordinatorInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Learning Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CoordinatorInput {
  name: string;
  /** @nullable */
  phone?: string | null;
  zoneId: number;
}