# lib/api-zod/src/generated/types/verifyTempPinResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Learning Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface VerifyTempPinResult {
  success: boolean;
  email: string;
}